+++ packages/com-bridge/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.0.27]
+### Fixed
+ - force com-brigge instance to be destroyed by JS engine in the same thread it's created (JSCLASS_FOREGROUND_FINALIZE)
+
 ##  [1.0.7]
 ### Fixed
  - Fix a @unitybas/com-bridge for SpiderMonkey52