+++ packages/mailer/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.1.17]
+### Fixed
+ - force mailer instance to be destroyed by JS engine in the same thread it's created (JSCLASS_FOREGROUND_FINALIZE)
+
 ## [5.1.16]
 ### Fixed
  - prevent expose a package to client by adding `"config": {"ubmodel": {} }` into package.json